refactor: update order-related models and services for consistency and clarity

+++ src/api/models/ingredient.ts
@@ -13,6 +13,7 @@ export class Ingredient implements IIngredient {
 
   @Column({ type: 'varchar', length: 255 })
   name!: string;
-  @Column({ type: 'float', precision: 10, scale: 2 })
+  
+  @Column({ type: 'numeric', precision: 10, scale: 2 })
   price!: number;
 }
