+++ src/api/models/product.ts
@@ -21,7 +21,7 @@ export class Product implements IProduct {
   @Column({ type: 'varchar', length: 255 })
   name!: string;
 
-  @Column({ type: 'float', precision: 10, scale: 2 })
+  @Column({ type: 'numeric', precision: 10, scale: 2 })
   price!: number;
 
   @OneToMany(() => Recipe, (recipe) => recipe.product)
